+++ packages/dara-core/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-  Added the ability to handle chunked responses to internal dara messages used by the `WebsocketHandler.send_and_wait` method. The messages will be gathered into a list before being returned to the caller.
+
 ## 1.7.4
 
 -   Added the ability to pass an asynchronous function to `ConfigurationBuilder.on_startup(...)`.